Inmate Views added
Now when the status of the inmate lists are captured, the user can click in and see individual view controllers for the inmate. The logic is in place to show it working, but as we get information and more models and inmates given, we can keep updating and adding to it.