LAMP Stack Implemented
LAMP Stack allows client to go to a website and add ghosts to a list. The iPad then could be connected to wifi and download the list of ghosts to be used by the application. The code gets the JSON data from the url.
Still need to implement translation between JSON to specific ghost models and specific gps location which will be an enumeration that is hard coded in the app.