Ghost Model Updated, Test Function Removed
The ghost model now has enums setup that take in the json strings are translates them into data needed for the ghosts such as coordinates, and 3d model files names.

The testing functions were removed from the app because the pins were hard coded in and they are now dynamic based on the ghost models.

Going forward we need to add the warning screen, start screen, and place a button on the start screen that allows our client to read in new JSON data.

We also need to collect information such as GPS locations for the ghosts, names, create new 3D models, etc.