Ghost Animations Modified.
Found a workflow for adding models with animations and animating based on whether or not they are captured and changing the animation when they are tapped.

NOTE: This currently is only setup for a ghost named "Snowden" and with a .scn file of snowden.scn. The rest still needs implemented.